--- a/fuzz_test.go
+++ b/fuzz_test.go
@@ -25,12 +25,25 @@ func FuzzParse(f *testing.F) {
 
 var benchFormulas = []string{"=0", "=SUM(A3+B9*2)/2"}
 
-func BenchmarkParse(b *testing.B) {
+func BenchmarkExcelParser(b *testing.B) {
 	for _, formula := range benchFormulas {
+		b.ResetTimer()
+
 		ps := efp.ExcelParser()
 		b.Run(formula, func(b *testing.B) {
 			for n := 0; n < b.N; n++ {
 				ps.Parse(formula)
+			}
+		})
+
+		b.Run(formula+`.PrettyPrint()`, func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				ps.PrettyPrint()
+			}
+		})
+		b.Run(formula+`.Render()`, func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				ps.Render()
 			}
 		})
 	}