# PR Details
Adds a benchmark and makes speed performance improvements

## Description

Command: `go test -short -bench . -benchmem`
Comparison (using benchcmp):
```
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                 old ns/op     new ns/op     delta
BenchmarkExcelParser/=0-10                                619399        313067        -49.46%
BenchmarkExcelParser/=0.PrettyPrint()-10                  146911363     1350755       -99.08%
BenchmarkExcelParser/=0.Render()-10                       9302861       203005        -97.82%
BenchmarkExcelParser/=SUM(A3+B9*2)/2-10                   506181        358507        -29.17%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.PrettyPrint()-10     100794675     1568738       -98.44%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.Render()-10          13432806      206847        -98.46%

benchmark                                                 old allocs     new allocs     delta
BenchmarkExcelParser/=0-10                                31             2              -93.55%
BenchmarkExcelParser/=0.PrettyPrint()-10                  10114          34             -99.66%
BenchmarkExcelParser/=0.Render()-10                       10100          21             -99.79%
BenchmarkExcelParser/=SUM(A3+B9*2)/2-10                   31             2              -93.55%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.PrettyPrint()-10     10122          35             -99.65%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.Render()-10          10109          21             -99.79%

benchmark                                                 old bytes      new bytes     delta
BenchmarkExcelParser/=0-10                                2033207        1406142       -30.84%
BenchmarkExcelParser/=0.PrettyPrint()-10                  1107912627     8394175       -99.24%
BenchmarkExcelParser/=0.Render()-10                       54198201       211704        -99.61%
BenchmarkExcelParser/=SUM(A3+B9*2)/2-10                   2036515        1419416       -30.30%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.PrettyPrint()-10     1109795994     10589631      -99.05%
BenchmarkExcelParser/=SUM(A3+B9*2)/2.Render()-10          54331288       211704        -99.61%
```

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.